+++ src/Cache/Manager.php
@@ -1,11 +1,13 @@
 <?php
 
+declare(strict_types=1);
+
 namespace Beebmx\KirbScheduler\Cache;
 
 use Illuminate\Cache\CacheManager;
 use Illuminate\Container\Container;
 
-class Manager extends CacheManager
+final class Manager extends CacheManager
 {
     public function __construct(Container $app)
     {