+++ src/Console/ScheduleRunCommand.php
@@ -7,7 +7,7 @@ namespace Beebmx\KirbScheduler\Console;
 use Beebmx\KirbScheduler\Facades\Schedule;
 use Kirby\CLI\CLI;
 
-class ScheduleRunCommand
+final class ScheduleRunCommand
 {
     public function __invoke(CLI $cli): void
     {