+++ tests/Feature/ArquitectureTest.php
@@ -1,5 +1,12 @@
 <?php
 
+arch()->preset()->php();
+
+arch()->preset()->security();
+
+arch()->preset()->strict()
+    ->ignoring('usleep');
+
 arch('globals')
     ->expect(['dd', 'dump', 'ray'])
     ->not->toBeUsed();