+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 All notable changes to `Scheduler` will be documented in this file.
 
+## 1.0.4 - 2025-11-05
+
+- Added `terminating` and `terminate` methods to Application class
+- Added strict types declaration
+
 ## 1.0.3 - 2025-10-14
 
 - Fixed an issue with base_path